+++ contracts/pools/LidoTempusPool.sol
@@ -66,17 +66,17 @@ contract LidoTempusPool is TempusPool {
 
     /// NOTE: Lido StETH is pegged 1:1 to ETH
     /// @return Asset Token amount
-    function numAssetsPerYieldToken(uint yieldTokens, uint) public pure override returns (uint) {
+    function numAssetsPerYieldToken(uint256 yieldTokens, uint256) public pure override returns (uint256) {
         return yieldTokens;
     }
 
     /// NOTE: Lido StETH is pegged 1:1 to ETH
     /// @return YBT amount
-    function numYieldTokensPerAsset(uint backingTokens, uint) public pure override returns (uint) {
+    function numYieldTokensPerAsset(uint256 backingTokens, uint256) public pure override returns (uint256) {
         return backingTokens;
     }
 
-    function interestRateToSharePrice(uint interestRate) internal pure override returns (uint) {
+    function interestRateToSharePrice(uint256 interestRate) internal pure override returns (uint256) {
         return interestRate; // no conversion needed, praise ETH
     }
 }
